+++ gen/cpp/Pythia8Extra.h
@@ -1,0 +1,4 @@
+// BinaryBuilder uses old SDK which doesn't have ranges
+#ifdef __APPLE__
+    #undef JLCXX_HAS_RANGES
+#endif